+++ releasenotes.md
@@ -1,2 +1,3 @@
 Whats New in v6.5
 ----------------------
+- You can now merge similar ingredients in formula view to one and sum up their quantities